Fix a truncation, found by pcc. Set a deliberately blank YFLAGS in Makefile, to override operating systems that set YFLAGS by default, like OpenBSD.

+++ iburg.c
@@ -198,7 +198,7 @@ static void *
 install(char *name)
 {
 	struct entry *p = alloc(sizeof *p);
-	int i = hash(name)%HASHSIZE;
+	long i = hash(name)%HASHSIZE;
 
 	p->sym.name = name;
 	p->link = table[i];
